requirements.txt
# Hugging Face API项目依赖

# 主要依赖
openai>=1.0.0
python-dotenv>=1.0.0

# 可选依赖（用于Jupyter笔记本等）
# jupyter>=1.0.0
# ipython>=8.0.0